+++ sync-for-expenses/CodatSyncExpenses/CodatSyncExpensesSDK.cs
@@ -121,10 +121,10 @@ namespace CodatSyncExpenses
         };
 
         private const string _language = "csharp";
-        private const string _sdkVersion = "3.1.0";
-        private const string _sdkGenVersion = "2.146.1";
+        private const string _sdkVersion = "3.2.0";
+        private const string _sdkGenVersion = "2.159.2";
         private const string _openapiDocVersion = "prealpha";
-        private const string _userAgent = "speakeasy-sdk/csharp 3.1.0 2.146.1 prealpha Codat.Sync.Expenses";
+        private const string _userAgent = "speakeasy-sdk/csharp 3.2.0 2.159.2 prealpha Codat.Sync.Expenses";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;