+++ sync-for-expenses/CodatSyncExpenses/Expenses.cs
@@ -60,10 +60,10 @@ namespace CodatSyncExpenses
     {
         public SDKConfig Config { get; private set; }
         private const string _language = "csharp";
-        private const string _sdkVersion = "3.1.0";
-        private const string _sdkGenVersion = "2.146.1";
+        private const string _sdkVersion = "3.2.0";
+        private const string _sdkGenVersion = "2.159.2";
         private const string _openapiDocVersion = "prealpha";
-        private const string _userAgent = "speakeasy-sdk/csharp 3.1.0 2.146.1 prealpha Codat.Sync.Expenses";
+        private const string _userAgent = "speakeasy-sdk/csharp 3.2.0 2.159.2 prealpha Codat.Sync.Expenses";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;
@@ -108,6 +108,7 @@ namespace CodatSyncExpenses
                 ContentType = contentType,
                 RawResponse = httpResponse
             };
+            
             if((response.StatusCode == 200))
             {
                 if(Utilities.IsContentTypeMatch("application/json",response.ContentType))
@@ -161,6 +162,7 @@ namespace CodatSyncExpenses
                 ContentType = contentType,
                 RawResponse = httpResponse
             };
+            
             if((response.StatusCode == 202))
             {
                 if(Utilities.IsContentTypeMatch("application/json",response.ContentType))
@@ -214,6 +216,7 @@ namespace CodatSyncExpenses
                 ContentType = contentType,
                 RawResponse = httpResponse
             };
+            
             if((response.StatusCode == 200))
             {
                 if(Utilities.IsContentTypeMatch("application/json",response.ContentType))